refactor: add directory to manage jinja templates
