--- src/create_github_project/resource_manager/__init__.py
@@ -1,5 +1,0 @@
-from ._resource_manager import ResourceManager
-
-__all__ = [
-    'ResourceManager'
-]
