--- src/create_github_project/resource_manager/resources/core/release/.versionrc.json
@@ -1,37 +1,0 @@
-{
-    "header": "# Changelog\n\nAll notable changes to this project will be documented in this file.",
-    "types": [
-        {
-            "type": "feat",
-            "section": "Features"
-        },
-        {
-            "type": "fix",
-            "section": "Bug Fixes"
-        },
-        {
-            "type": "test",
-            "section": "Tests"
-        },
-        {
-            "type": "docs",
-            "section": "Documentation"
-        },
-        {
-            "type": "style",
-            "section": "Styles"
-        },
-        {
-            "type": "refactor",
-            "section": "Refactor"
-        },
-        {
-            "type": "perf",
-            "section": "Performance"
-        },
-        {
-            "type": "chore",
-            "section": "Chore"
-        }
-    ]
-}
