+++ src/create_github_project/templates/_core/.github/ISSUE_TEMPLATE/2_bug.md
