+++ src/create_github_project/templates/_core/.github/PULL_REQUEST_TEMPLATE.md
