+++ src/create_github_project/templates/_core/.github/workflows/EXCLUDE-release-request.yml.jinja
