+++ src/create_github_project/templates/_core/.github/workflows/do-release.yml.jinja
