+++ src/create_github_project/templates/_core/.github/workflows/major-release-request.yml.jinja
@@ -1,0 +1,3 @@
+{% with release_stage="major", production_branch=production_branch %}
+{% include '_core/.github/workflows/EXCLUDE-release-request.yml.jinja' %}
+{% endwith %}
