+++ src/create_github_project/templates/_core/.github/workflows/minor-release-request.yml.jinja
@@ -1,0 +1,3 @@
+{% with release_stage="minor", production_branch=production_branch %}
+{% include '_core/.github/workflows/EXCLUDE-release-request.yml.jinja' %}
+{% endwith %}
