+++ src/create_github_project/templates/_core/.gitignore
