+++ src/create_github_project/templates/_core/release/package.json
