+++ src/create_github_project/templates/_extension/java/docs/setup/schema.xml
