+++ src/create_github_project/templates/_extension/python/docs/setup/python.md
