+++ src/create_github_project/templates/default/README.md.jinja
